test: 개발을 위한 데이터 테스트와 서비스 로직 테스트 분리
